% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testForwardContext.R
\docType{package}
\name{testForwardContext-package}
\alias{testForwardContext}
\alias{testForwardContext-package}
\title{testForwardContext: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
